- Check gs_data_dir raster data subfolders
(cherry picked from commit 541467b197439d7817e618880170f69b4fb16439)

+++ geonode/br/management/commands/backup.py
@@ -377,16 +377,35 @@ class Command(BaseCommand):
     def dump_geoserver_raster_data(self, config, settings, target_folder):
         if (config.gs_data_dir):
             if (config.gs_dump_raster_data):
-                # Dump '$config.gs_data_dir/data/geonode'
+                # Dump '$config.gs_data_dir/geonode'
                 gs_data_root = os.path.join(config.gs_data_dir, 'geonode')
                 if not os.path.isabs(gs_data_root):
                     gs_data_root = os.path.join(settings.PROJECT_ROOT, '..', gs_data_root)
-                gs_data_folder = os.path.join(target_folder, 'gs_data_dir', 'geonode')
-                if not os.path.exists(gs_data_folder):
-                    os.makedirs(gs_data_folder)
+                print("Dumping GeoServer Uploaded Data from '"+gs_data_root+"'.")
+                if os.path.exists(gs_data_root):
+                    gs_data_folder = os.path.join(target_folder, 'gs_data_dir', 'geonode')
+                    if not os.path.exists(gs_data_folder):
+                        os.makedirs(gs_data_folder)
 
-                copy_tree(gs_data_root, gs_data_folder)
-                print("Dumped GeoServer Uploaded Data from '"+gs_data_root+"'.")
+                    copy_tree(gs_data_root, gs_data_folder)
+                    print("Dumped GeoServer Uploaded Data from '"+gs_data_root+"'.")
+                else:
+                    print("Skipped GeoServer Uploaded Data '"+gs_data_root+"'.")
+
+                # Dump '$config.gs_data_dir/data/geonode'
+                gs_data_root = os.path.join(config.gs_data_dir, 'data', 'geonode')
+                if not os.path.isabs(gs_data_root):
+                    gs_data_root = os.path.join(settings.PROJECT_ROOT, '..', gs_data_root)
+                print("Dumping GeoServer Uploaded Data from '"+gs_data_root+"'.")
+                if os.path.exists(gs_data_root):
+                    gs_data_folder = os.path.join(target_folder, 'gs_data_dir', 'data', 'geonode')
+                    if not os.path.exists(gs_data_folder):
+                        os.makedirs(gs_data_folder)
+
+                    copy_tree(gs_data_root, gs_data_folder)
+                    print("Dumped GeoServer Uploaded Data from '" + gs_data_root + "'.")
+                else:
+                    print("Skipped GeoServer Uploaded Data '"+gs_data_root+"'.")
 
     def dump_geoserver_vector_data(self, config, settings, target_folder):
         if (config.gs_dump_vector_data):
@@ -429,22 +448,30 @@ class Command(BaseCommand):
                 match = regexp.match(text)
                 if match:
                     relpath = match.group(1)
-                    abspath = relpath if os.path.isabs(relpath) else \
-                        os.path.abspath(
-                            os.path.join(os.path.dirname(path), relpath))
-                    if os.path.exists(abspath):
-                        return abspath
+                    try:
+                        abspath = relpath if os.path.isabs(relpath) else \
+                            os.path.abspath(
+                                os.path.join(os.path.dirname(path), relpath))
+                        if os.path.exists(abspath):
+                            return abspath
+                    except Exception:
+                        print("WARNING: Error while trying to dump {}".format(text))
+                        return
 
         def match_fileurl(key, text, regexp=re.compile("^file:(.+)$")):
             if key in ('url', ):
                 match = regexp.match(text)
                 if match:
                     relpath = match.group(1)
-                    abspath = relpath if os.path.isabs(relpath) else \
-                        os.path.abspath(
-                            os.path.join(config.gs_data_dir, relpath))
-                    if os.path.exists(abspath):
-                        return abspath
+                    try:
+                        abspath = relpath if os.path.isabs(relpath) else \
+                            os.path.abspath(
+                                os.path.join(config.gs_data_dir, relpath))
+                        if os.path.exists(abspath):
+                            return abspath
+                    except Exception:
+                        print("WARNING: Error while trying to dump {}".format(text))
+                        return
 
         def dump_external_resources_from_xml(path):
